+++ tests/helpers/command-generator.js
