+++ tests/helpers/package-cache.js
